infranest/core/requirements.txt
# Core Dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Jinja2==3.1.2
PyYAML==6.0.1
requests==2.31.0
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Job Queue
Celery==5.3.4
redis==5.0.1

# AI Integration
openai==1.3.7
anthropic==0.8.1

# Machine Learning and LLM (Optional - only for local models)
# torch>=2.0.0
# transformers>=4.30.0
# accelerate>=0.20.0
# bitsandbytes>=0.39.0

# GGUF Model Support (Not needed - we use Groq API)
# llama-cpp-python>=0.2.0

# Code Generation
black==23.11.0
isort==5.12.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
python-decouple==3.8

# Evaluation & Benchmarking (PRISM Step 9)
matplotlib==3.8.2
numpy==1.26.2
scipy==1.11.4

# Code Analysis (PRISM Refiner)
pylint==3.0.3
autopep8==2.0.4